requirements.txt
ninja
opencv-python-headless==4.5.5.64
tensorboardX
tqdm
matplotlib
scipy
braceexpand
imageio
imageio-ffmpeg
omegaconf
ffmpegcv
mediapy
OpenEXR
pillow>=10.3.0
diffusers==0.32.0
accelerate==1.3.0
transformers==4.47.1
peft==0.7.0
datasets==2.19.0